requirements.txt
sflask
flask-cors
firebase-admin
python-dotenv